chore: move text.py
